mesh/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
Booktype==1.5
click==8.2.1
Django==5.2.1
dnspython==2.7.0
ephem==4.2
fastapi==0.115.12
h11==0.16.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
nats-py==2.10.0
numpy==2.2.6
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
pymongo==4.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==4.5.5
scipy==1.15.3
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.34.2
